requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.0
python-decouple==3.6
router==0.1
sniffio==1.2.0
starlette==0.20.0
typing_extensions==4.2.0
uvicorn==0.17.6
pytest==7.1.2
jinja2==3.1.2